+++ p3_datawarehouse_aws_redshift/src/createenv/create_tables.py
@@ -1,6 +1,5 @@
-import configparser
 import psycopg2
-from p3_datawarehouse_aws_redshift.src.sql_queries import create_table_queries, drop_table_queries
+from p3_datawarehouse_aws_redshift.src.etl.sql_queries_etl import create_table_queries, drop_table_queries
 
 def drop_tables(cur, conn):
     for query in drop_table_queries:
