+++ p3_datawarehouse_aws_redshift/src/etl/sql_queries_etl.py
